--- app/models/levelling_up_premium_payments/award.rb
@@ -1,2807 +1,0 @@
-require "csv"
-
-module LevellingUpPremiumPayments
-  class Award < ApplicationRecord
-    self.table_name = "levelling_up_premium_payments_awards"
-
-    belongs_to :school, foreign_key: :school_urn, primary_key: :urn, inverse_of: :levelling_up_premium_payments_awards, optional: true
-
-    validates :academic_year, presence: true
-    validates :school_urn, presence: true, numericality: true
-    validates :award_amount, presence: true
-    validates :award_amount, numericality: {greater_than: 0}
-
-    def self.csv_for_academic_year(academic_year)
-      attribute_names = [:school_urn, :award_amount]
-
-      CSV.generate(headers: true) do |csv|
-        csv << attribute_names
-
-        where(academic_year: academic_year.to_s).each do |row|
-          csv << attribute_names.map { |attr| row.send(attr) }
-        end
-      end
-    end
-
-    def self.last_updated_at(academic_year)
-      where(academic_year: academic_year.to_s).last&.updated_at
-    end
-
-    def self.distinct_academic_years
-      select(:academic_year).distinct.order(academic_year: :desc).map(&:academic_year)
-    end
-
-    # TODO: this data is OUTDATED and the method should be removed some time after the database migration is run
-    private_class_method def self.urn_to_award_amount_in_pounds_for_2022_to_2023
-      @urn_to_award_amount_in_pounds_for_2022_to_2023 ||= {
-        100006 => 2_000,
-        100049 => 2_000,
-        100050 => 2_000,
-        100051 => 2_000,
-        100052 => 2_000,
-        100053 => 1_500,
-        100054 => 1_500,
-        100055 => 2_000,
-        100056 => 2_000,
-        100059 => 2_000,
-        100060 => 2_000,
-        100092 => 2_000,
-        100094 => 2_000,
-        100096 => 2_000,
-        100103 => 2_000,
-        100171 => 2_000,
-        100183 => 1_500,
-        100190 => 1_500,
-        100204 => 2_000,
-        100277 => 2_000,
-        100279 => 1_500,
-        100282 => 2_000,
-        100284 => 2_000,
-        100285 => 2_000,
-        100307 => 2_000,
-        100311 => 2_000,
-        100312 => 2_000,
-        100379 => 2_000,
-        100381 => 2_000,
-        100382 => 2_000,
-        100391 => 2_000,
-        100453 => 2_000,
-        100455 => 2_000,
-        100457 => 2_000,
-        100458 => 2_000,
-        100459 => 2_000,
-        100467 => 2_000,
-        100469 => 2_000,
-        100502 => 1_500,
-        100503 => 2_000,
-        100549 => 2_000,
-        100590 => 1_500,
-        100624 => 1_500,
-        100625 => 2_000,
-        100627 => 2_000,
-        100637 => 2_000,
-        100642 => 2_000,
-        100643 => 2_000,
-        100654 => 2_000,
-        100659 => 2_000,
-        100740 => 2_000,
-        100742 => 2_000,
-        100745 => 1_500,
-        100747 => 1_500,
-        100748 => 2_000,
-        100749 => 1_500,
-        100760 => 2_000,
-        100763 => 2_000,
-        100765 => 2_000,
-        100849 => 2_000,
-        100857 => 2_000,
-        100859 => 2_000,
-        100872 => 2_000,
-        100875 => 2_000,
-        100876 => 2_000,
-        100880 => 2_000,
-        100889 => 2_000,
-        100965 => 2_000,
-        100966 => 2_000,
-        100967 => 2_000,
-        100972 => 2_000,
-        100973 => 2_000,
-        100974 => 2_000,
-        100975 => 2_000,
-        100977 => 2_000,
-        100978 => 2_000,
-        100986 => 2_000,
-        100987 => 2_000,
-        100989 => 2_000,
-        100994 => 2_000,
-        101053 => 1_500,
-        101094 => 2_000,
-        101102 => 2_000,
-        101103 => 2_000,
-        101154 => 2_000,
-        101182 => 2_000,
-        101184 => 2_000,
-        101243 => 1_500,
-        101244 => 1_500,
-        101245 => 1_500,
-        101255 => 2_000,
-        101345 => 1_500,
-        101397 => 2_000,
-        101487 => 2_000,
-        101493 => 2_000,
-        101564 => 1_500,
-        101697 => 1_500,
-        101706 => 2_000,
-        101811 => 1_500,
-        101814 => 2_000,
-        101821 => 1_500,
-        101851 => 2_000,
-        101852 => 2_000,
-        101856 => 2_000,
-        101928 => 1_500,
-        101939 => 1_500,
-        101940 => 1_500,
-        101941 => 1_500,
-        101943 => 1_500,
-        101965 => 2_000,
-        101969 => 2_000,
-        101970 => 2_000,
-        101971 => 1_500,
-        101972 => 2_000,
-        102049 => 1_500,
-        102052 => 2_000,
-        102056 => 1_500,
-        102066 => 2_000,
-        102067 => 2_000,
-        102069 => 2_000,
-        102070 => 2_000,
-        102153 => 1_500,
-        102157 => 2_000,
-        102175 => 2_000,
-        102176 => 2_000,
-        102178 => 2_000,
-        102180 => 2_000,
-        102260 => 1_500,
-        102362 => 1_500,
-        102449 => 1_500,
-        102451 => 1_500,
-        102462 => 2_000,
-        102464 => 1_500,
-        102539 => 1_500,
-        102554 => 2_000,
-        102555 => 2_000,
-        102562 => 2_000,
-        102674 => 1_500,
-        102697 => 2_000,
-        102698 => 2_000,
-        102699 => 2_000,
-        102708 => 2_000,
-        102776 => 1_500,
-        102782 => 1_500,
-        102786 => 1_500,
-        102787 => 1_500,
-        102794 => 2_000,
-        102849 => 1_500,
-        102858 => 1_500,
-        102878 => 1_500,
-        102879 => 2_000,
-        103025 => 2_000,
-        103080 => 1_500,
-        103094 => 1_500,
-        103097 => 1_500,
-        103100 => 1_500,
-        103105 => 1_500,
-        103106 => 1_500,
-        103119 => 1_500,
-        103146 => 2_000,
-        103483 => 2_000,
-        103486 => 2_000,
-        103493 => 2_000,
-        103497 => 1_500,
-        103498 => 2_000,
-        103501 => 1_500,
-        103503 => 2_000,
-        103509 => 2_000,
-        103514 => 2_000,
-        103519 => 2_000,
-        103534 => 2_000,
-        103539 => 2_000,
-        103560 => 1_500,
-        103562 => 1_500,
-        103563 => 2_000,
-        103601 => 2_000,
-        103603 => 2_000,
-        103605 => 2_000,
-        103606 => 2_000,
-        103611 => 2_000,
-        103613 => 2_000,
-        103614 => 2_000,
-        103622 => 2_000,
-        103625 => 2_000,
-        103628 => 2_000,
-        103632 => 2_000,
-        103760 => 2_500,
-        103765 => 3_000,
-        103855 => 2_000,
-        103858 => 2_500,
-        103877 => 3_000,
-        103878 => 3_000,
-        103880 => 3_000,
-        103881 => 3_000,
-        103882 => 3_000,
-        103883 => 2_500,
-        103887 => 3_000,
-        104018 => 2_500,
-        104019 => 2_500,
-        104038 => 2_000,
-        104130 => 1_500,
-        104131 => 1_500,
-        104132 => 2_000,
-        104133 => 2_000,
-        104255 => 2_000,
-        104259 => 2_500,
-        104269 => 3_000,
-        104272 => 3_000,
-        104288 => 2_000,
-        104387 => 2_000,
-        104395 => 2_000,
-        104412 => 2_000,
-        104414 => 2_000,
-        104415 => 2_000,
-        104417 => 2_000,
-        104418 => 3_000,
-        104495 => 3_000,
-        104498 => 3_000,
-        104688 => 3_000,
-        104698 => 2_500,
-        104700 => 3_000,
-        104705 => 2_000,
-        104706 => 3_000,
-        104712 => 2_500,
-        104713 => 2_500,
-        104714 => 2_500,
-        104715 => 3_000,
-        104717 => 2_500,
-        104734 => 3_000,
-        104735 => 3_000,
-        104736 => 3_000,
-        104739 => 3_000,
-        104742 => 3_000,
-        104744 => 3_000,
-        104748 => 3_000,
-        104749 => 3_000,
-        104829 => 2_500,
-        104835 => 3_000,
-        104843 => 3_000,
-        104849 => 3_000,
-        104956 => 2_000,
-        104959 => 3_000,
-        104960 => 2_000,
-        104962 => 2_000,
-        104964 => 2_000,
-        104977 => 3_000,
-        104979 => 2_500,
-        104980 => 3_000,
-        104982 => 3_000,
-        104983 => 3_000,
-        105097 => 3_000,
-        105103 => 2_500,
-        105107 => 2_500,
-        105129 => 3_000,
-        105130 => 3_000,
-        105131 => 2_500,
-        105133 => 3_000,
-        105135 => 1_500,
-        105137 => 3_000,
-        105139 => 3_000,
-        105252 => 2_000,
-        105253 => 2_000,
-        105262 => 2_000,
-        105263 => 2_500,
-        105264 => 2_000,
-        105276 => 3_000,
-        105277 => 3_000,
-        105280 => 3_000,
-        105354 => 2_000,
-        105355 => 2_500,
-        105358 => 2_000,
-        105360 => 2_500,
-        105376 => 3_000,
-        105560 => 3_000,
-        105574 => 3_000,
-        105576 => 3_000,
-        105581 => 3_000,
-        105602 => 3_000,
-        105608 => 3_000,
-        105614 => 3_000,
-        105623 => 3_000,
-        105736 => 2_000,
-        105738 => 2_500,
-        105837 => 2_500,
-        105844 => 2_000,
-        105845 => 3_000,
-        105861 => 3_000,
-        105986 => 2_000,
-        105989 => 2_000,
-        106022 => 2_000,
-        106023 => 2_000,
-        106166 => 2_000,
-        106167 => 2_000,
-        106172 => 1_500,
-        106173 => 1_500,
-        106266 => 2_000,
-        106268 => 2_500,
-        106271 => 2_000,
-        106278 => 3_000,
-        106279 => 3_000,
-        106365 => 1_500,
-        106370 => 1_500,
-        106375 => 1_500,
-        106391 => 1_500,
-        106523 => 1_500,
-        106535 => 1_500,
-        106543 => 2_000,
-        106666 => 3_000,
-        106818 => 3_000,
-        106962 => 2_000,
-        106966 => 3_000,
-        106970 => 3_000,
-        107140 => 1_500,
-        107169 => 2_000,
-        107178 => 2_000,
-        107395 => 2_000,
-        107588 => 2_000,
-        107589 => 2_000,
-        107756 => 3_000,
-        107775 => 3_000,
-        107778 => 2_500,
-        107782 => 2_500,
-        107796 => 3_000,
-        107797 => 2_500,
-        107799 => 3_000,
-        107801 => 3_000,
-        107802 => 3_000,
-        108055 => 3_000,
-        108057 => 2_000,
-        108058 => 2_500,
-        108059 => 2_500,
-        108075 => 2_500,
-        108079 => 2_000,
-        108088 => 2_000,
-        108097 => 3_000,
-        108119 => 3_000,
-        108120 => 2_500,
-        108123 => 3_000,
-        108133 => 2_500,
-        108139 => 3_000,
-        108311 => 3_000,
-        108410 => 2_000,
-        108426 => 2_000,
-        108551 => 2_000,
-        108565 => 2_000,
-        108628 => 2_000,
-        108637 => 1_500,
-        108640 => 1_500,
-        108641 => 1_500,
-        108642 => 1_500,
-        108644 => 2_000,
-        108645 => 1_500,
-        108652 => 1_500,
-        108653 => 2_000,
-        108657 => 1_500,
-        108666 => 3_000,
-        108727 => 3_000,
-        108730 => 3_000,
-        108731 => 3_000,
-        108738 => 3_000,
-        108741 => 3_000,
-        109385 => 2_000,
-        109386 => 2_000,
-        109391 => 1_500,
-        109393 => 2_000,
-        109403 => 2_500,
-        109406 => 2_500,
-        109407 => 2_500,
-        109409 => 2_000,
-        109410 => 2_000,
-        109661 => 2_000,
-        109669 => 2_000,
-        109686 => 2_500,
-        109690 => 2_500,
-        109694 => 2_000,
-        109707 => 2_500,
-        109709 => 2_500,
-        109739 => 2_500,
-        109742 => 2_500,
-        109744 => 3_000,
-        109746 => 2_500,
-        110078 => 1_500,
-        110180 => 1_500,
-        110181 => 2_000,
-        110182 => 1_500,
-        110183 => 1_500,
-        110186 => 1_500,
-        110190 => 1_500,
-        110532 => 1_500,
-        110575 => 1_500,
-        110576 => 1_500,
-        110579 => 1_500,
-        110581 => 2_000,
-        110582 => 1_500,
-        110584 => 1_500,
-        110587 => 2_000,
-        110592 => 1_500,
-        110882 => 2_500,
-        110907 => 2_000,
-        110943 => 3_000,
-        110948 => 3_000,
-        110951 => 2_500,
-        111396 => 2_000,
-        111417 => 1_500,
-        111451 => 1_500,
-        111457 => 3_000,
-        111494 => 1_500,
-        111495 => 2_000,
-        111496 => 1_500,
-        111498 => 2_000,
-        111499 => 2_000,
-        111501 => 2_000,
-        111503 => 2_000,
-        111504 => 1_500,
-        111505 => 2_000,
-        111507 => 1_500,
-        111508 => 1_500,
-        111511 => 2_000,
-        111514 => 3_000,
-        111515 => 3_000,
-        111517 => 2_000,
-        111748 => 2_000,
-        111773 => 3_000,
-        111777 => 2_000,
-        112041 => 2_000,
-        112045 => 2_000,
-        112054 => 2_000,
-        112067 => 2_500,
-        112095 => 2_000,
-        112096 => 2_000,
-        112098 => 2_000,
-        112375 => 2_000,
-        112377 => 2_000,
-        112382 => 1_500,
-        112399 => 1_500,
-        112464 => 2_000,
-        112466 => 1_500,
-        112939 => 3_000,
-        112949 => 2_000,
-        112951 => 3_000,
-        112958 => 3_000,
-        112961 => 3_000,
-        112970 => 2_000,
-        112991 => 2_500,
-        113031 => 3_000,
-        113033 => 3_000,
-        113040 => 3_000,
-        113044 => 3_000,
-        113526 => 1_500,
-        113533 => 3_000,
-        113551 => 1_500,
-        113633 => 2_000,
-        113634 => 1_500,
-        113635 => 1_500,
-        113636 => 2_000,
-        113637 => 2_000,
-        113638 => 2_000,
-        113641 => 2_000,
-        113643 => 2_000,
-        113644 => 3_000,
-        113645 => 3_000,
-        113648 => 3_000,
-        113649 => 3_000,
-        113650 => 3_000,
-        113651 => 3_000,
-        113654 => 2_000,
-        113854 => 2_000,
-        113861 => 3_000,
-        113863 => 2_000,
-        113875 => 2_000,
-        113942 => 2_000,
-        113954 => 2_000,
-        113955 => 2_000,
-        113956 => 2_000,
-        113957 => 2_500,
-        113960 => 2_500,
-        113961 => 2_000,
-        113965 => 2_500,
-        113992 => 3_000,
-        114305 => 3_000,
-        114308 => 2_500,
-        114311 => 3_000,
-        114315 => 3_000,
-        114337 => 3_000,
-        114340 => 3_000,
-        114345 => 3_000,
-        114349 => 3_000,
-        114581 => 1_500,
-        114588 => 2_000,
-        114607 => 1_500,
-        114619 => 2_000,
-        114677 => 2_500,
-        114678 => 2_000,
-        114680 => 2_000,
-        114687 => 1_500,
-        114688 => 2_000,
-        115237 => 1_500,
-        115457 => 1_500,
-        115464 => 1_500,
-        115466 => 2_000,
-        115469 => 1_500,
-        115471 => 2_000,
-        115475 => 2_000,
-        115821 => 1_500,
-        115823 => 2_000,
-        115825 => 1_500,
-        115836 => 2_000,
-        115837 => 2_000,
-        115838 => 2_000,
-        115847 => 2_000,
-        116234 => 1_500,
-        116432 => 1_500,
-        116440 => 1_500,
-        116445 => 1_500,
-        116447 => 1_500,
-        116450 => 1_500,
-        116453 => 2_000,
-        116463 => 2_500,
-        116465 => 2_000,
-        116469 => 1_500,
-        116473 => 2_000,
-        116505 => 2_000,
-        116511 => 1_500,
-        116603 => 2_000,
-        116609 => 2_000,
-        116611 => 2_000,
-        116614 => 1_500,
-        116617 => 2_000,
-        116618 => 1_500,
-        116622 => 2_000,
-        116624 => 2_000,
-        116637 => 1_500,
-        116639 => 2_000,
-        116641 => 2_000,
-        116958 => 1_500,
-        117049 => 1_500,
-        117056 => 1_500,
-        117062 => 1_500,
-        117064 => 1_500,
-        117504 => 1_500,
-        117518 => 1_500,
-        117534 => 1_500,
-        117552 => 1_500,
-        117665 => 2_000,
-        117667 => 1_500,
-        117669 => 2_000,
-        117671 => 2_000,
-        117672 => 1_500,
-        117674 => 2_000,
-        117679 => 2_000,
-        117680 => 1_500,
-        117681 => 1_500,
-        117682 => 1_500,
-        117683 => 1_500,
-        117684 => 1_500,
-        117685 => 2_000,
-        117686 => 2_000,
-        117690 => 1_500,
-        118075 => 1_500,
-        118085 => 1_500,
-        118111 => 2_000,
-        118138 => 2_000,
-        118140 => 2_000,
-        118144 => 2_000,
-        118145 => 1_500,
-        118146 => 2_000,
-        118148 => 1_500,
-        118226 => 2_500,
-        118227 => 3_000,
-        118788 => 1_500,
-        118879 => 1_500,
-        118898 => 2_000,
-        118903 => 1_500,
-        118908 => 1_500,
-        118928 => 1_500,
-        119027 => 2_000,
-        119029 => 1_500,
-        119036 => 2_000,
-        119037 => 2_000,
-        119040 => 2_000,
-        119041 => 2_000,
-        119042 => 2_000,
-        119044 => 2_000,
-        119045 => 2_000,
-        119046 => 2_000,
-        119050 => 2_000,
-        119051 => 1_500,
-        119052 => 1_500,
-        119055 => 1_500,
-        119056 => 2_000,
-        119058 => 2_000,
-        119059 => 1_500,
-        119062 => 2_000,
-        119112 => 2_000,
-        119707 => 1_500,
-        119723 => 1_500,
-        119743 => 1_500,
-        119771 => 2_000,
-        119773 => 2_000,
-        119774 => 2_000,
-        119780 => 1_500,
-        119781 => 2_000,
-        119782 => 1_500,
-        119790 => 2_000,
-        119798 => 2_000,
-        119804 => 1_500,
-        119861 => 1_500,
-        119866 => 2_000,
-        119868 => 3_000,
-        119871 => 3_000,
-        119876 => 2_000,
-        119878 => 2_000,
-        119880 => 2_000,
-        119883 => 2_000,
-        119887 => 1_500,
-        119889 => 1_500,
-        119892 => 1_500,
-        119893 => 1_500,
-        119895 => 2_000,
-        119897 => 2_000,
-        119898 => 2_000,
-        120281 => 1_500,
-        120292 => 1_500,
-        120298 => 1_500,
-        120348 => 1_500,
-        120352 => 1_500,
-        120361 => 2_000,
-        120363 => 2_000,
-        120753 => 3_000,
-        120754 => 3_000,
-        120755 => 3_000,
-        121254 => 3_000,
-        121256 => 3_000,
-        121257 => 3_000,
-        121258 => 3_000,
-        121260 => 2_000,
-        121261 => 2_500,
-        121262 => 3_000,
-        121264 => 2_500,
-        121265 => 3_000,
-        121270 => 2_000,
-        121663 => 2_000,
-        121667 => 2_000,
-        121668 => 2_000,
-        121702 => 2_000,
-        121764 => 3_000,
-        121765 => 2_000,
-        121766 => 3_000,
-        121772 => 3_000,
-        121776 => 3_000,
-        121778 => 2_500,
-        121779 => 2_500,
-        122162 => 2_000,
-        122335 => 1_500,
-        122348 => 1_500,
-        122363 => 1_500,
-        122374 => 1_500,
-        122382 => 1_500,
-        122383 => 2_000,
-        122384 => 2_000,
-        122389 => 2_000,
-        122390 => 2_000,
-        122949 => 3_000,
-        122953 => 2_500,
-        122955 => 3_000,
-        122956 => 2_500,
-        122957 => 2_500,
-        122961 => 2_500,
-        122964 => 3_000,
-        123329 => 1_500,
-        123332 => 1_500,
-        123333 => 1_500,
-        123337 => 2_000,
-        123349 => 2_000,
-        123631 => 2_000,
-        123635 => 2_000,
-        123862 => 2_000,
-        123875 => 2_000,
-        123878 => 2_500,
-        123883 => 2_000,
-        123938 => 3_000,
-        123940 => 3_000,
-        123944 => 3_000,
-        123945 => 3_000,
-        124395 => 1_500,
-        124453 => 2_000,
-        124467 => 2_000,
-        124498 => 3_000,
-        124500 => 3_000,
-        124508 => 2_000,
-        124856 => 2_000,
-        124903 => 3_000,
-        125452 => 1_500,
-        125456 => 1_500,
-        125474 => 1_500,
-        125480 => 2_000,
-        125481 => 2_000,
-        125794 => 1_500,
-        125805 => 1_500,
-        126071 => 1_500,
-        126154 => 2_000,
-        126155 => 2_000,
-        126157 => 1_500,
-        126161 => 1_500,
-        126163 => 1_500,
-        126170 => 2_000,
-        126171 => 3_000,
-        126549 => 3_000,
-        126550 => 2_000,
-        126705 => 2_000,
-        126712 => 2_000,
-        127715 => 3_000,
-        128604 => 3_000,
-        128614 => 2_000,
-        128768 => 2_000,
-        128851 => 3_000,
-        128922 => 3_000,
-        128932 => 2_000,
-        128991 => 3_000,
-        129050 => 2_000,
-        129342 => 2_000,
-        130247 => 2_000,
-        130316 => 3_000,
-        130344 => 2_000,
-        130349 => 1_500,
-        130359 => 1_500,
-        130371 => 2_000,
-        130856 => 2_000,
-        130908 => 2_500,
-        130909 => 2_000,
-        130912 => 2_000,
-        130942 => 2_000,
-        130961 => 3_000,
-        130991 => 2_000,
-        130996 => 2_500,
-        131022 => 3_000,
-        131062 => 2_000,
-        131065 => 3_000,
-        131066 => 2_000,
-        131068 => 1_500,
-        131099 => 2_000,
-        131100 => 2_000,
-        131102 => 2_000,
-        131134 => 2_000,
-        131182 => 2_000,
-        131200 => 1_500,
-        131201 => 2_000,
-        131258 => 1_500,
-        131262 => 2_000,
-        131277 => 3_000,
-        131294 => 3_000,
-        131322 => 3_000,
-        131425 => 3_000,
-        131479 => 1_500,
-        131506 => 3_000,
-        131512 => 3_000,
-        131526 => 3_000,
-        131530 => 2_000,
-        131535 => 2_000,
-        131544 => 2_000,
-        131547 => 2_000,
-        131552 => 2_000,
-        131584 => 2_000,
-        131596 => 2_000,
-        131609 => 2_000,
-        131619 => 3_000,
-        131626 => 2_500,
-        131629 => 3_000,
-        131652 => 2_000,
-        131690 => 2_000,
-        131697 => 2_000,
-        131726 => 2_000,
-        131747 => 2_000,
-        131748 => 2_000,
-        131749 => 2_000,
-        131752 => 2_000,
-        131753 => 2_000,
-        131757 => 1_500,
-        131769 => 2_000,
-        131772 => 3_000,
-        131827 => 3_000,
-        131838 => 2_000,
-        131880 => 3_000,
-        131885 => 2_000,
-        131889 => 2_000,
-        131895 => 2_000,
-        131896 => 2_000,
-        131897 => 1_500,
-        131905 => 3_000,
-        131945 => 2_000,
-        131987 => 2_000,
-        132027 => 3_000,
-        132033 => 3_000,
-        132051 => 2_000,
-        132058 => 2_000,
-        132128 => 2_000,
-        132130 => 2_000,
-        132131 => 2_000,
-        132133 => 3_000,
-        132148 => 1_500,
-        132217 => 2_000,
-        132231 => 3_000,
-        132233 => 3_000,
-        132711 => 2_000,
-        132727 => 2_000,
-        132731 => 2_000,
-        132741 => 3_000,
-        132771 => 2_000,
-        132816 => 2_000,
-        132905 => 3_000,
-        132993 => 3_000,
-        133055 => 3_000,
-        133058 => 3_000,
-        133085 => 2_000,
-        133114 => 1_500,
-        133115 => 2_000,
-        133117 => 3_000,
-        133118 => 3_000,
-        133119 => 3_000,
-        133125 => 2_000,
-        133164 => 3_000,
-        133289 => 2_000,
-        133317 => 1_500,
-        133386 => 2_000,
-        133398 => 2_000,
-        133409 => 3_000,
-        133421 => 3_000,
-        133432 => 2_000,
-        133441 => 3_000,
-        133561 => 1_500,
-        133580 => 1_500,
-        133581 => 2_000,
-        133583 => 2_000,
-        133653 => 2_000,
-        133660 => 2_000,
-        133675 => 2_000,
-        133689 => 1_500,
-        133697 => 2_500,
-        133719 => 2_500,
-        133725 => 3_000,
-        133743 => 1_500,
-        133744 => 3_000,
-        133748 => 3_000,
-        133749 => 2_000,
-        133768 => 3_000,
-        133778 => 2_000,
-        133779 => 2_000,
-        134003 => 2_000,
-        134064 => 2_000,
-        134108 => 2_000,
-        134127 => 2_000,
-        134130 => 2_000,
-        134159 => 3_000,
-        134190 => 1_500,
-        134222 => 2_000,
-        134223 => 2_500,
-        134224 => 3_000,
-        134225 => 2_000,
-        134226 => 1_500,
-        134253 => 3_000,
-        134256 => 2_000,
-        134257 => 2_000,
-        134269 => 3_000,
-        134274 => 2_000,
-        134283 => 2_000,
-        134297 => 2_000,
-        134311 => 2_000,
-        134314 => 2_000,
-        134321 => 3_000,
-        134366 => 2_000,
-        134367 => 2_000,
-        134369 => 2_000,
-        134373 => 3_000,
-        134374 => 2_000,
-        134390 => 2_000,
-        134523 => 3_000,
-        134525 => 3_000,
-        134533 => 3_000,
-        134623 => 2_000,
-        134625 => 2_000,
-        134635 => 2_000,
-        134640 => 1_500,
-        134646 => 2_500,
-        134658 => 3_000,
-        134662 => 3_000,
-        134693 => 2_000,
-        134694 => 3_000,
-        134697 => 3_000,
-        134699 => 3_000,
-        134727 => 1_500,
-        134758 => 3_000,
-        134759 => 3_000,
-        134765 => 3_000,
-        134768 => 3_000,
-        134783 => 1_500,
-        134798 => 2_000,
-        134813 => 3_000,
-        134814 => 1_500,
-        134815 => 2_000,
-        134846 => 2_000,
-        134859 => 2_000,
-        134865 => 3_000,
-        134872 => 2_000,
-        134880 => 2_000,
-        134884 => 3_000,
-        134888 => 2_500,
-        134889 => 2_500,
-        134919 => 2_000,
-        134937 => 2_500,
-        134970 => 2_000,
-        134971 => 1_500,
-        134972 => 3_000,
-        134993 => 2_000,
-        134994 => 2_000,
-        134996 => 1_500,
-        135003 => 1_500,
-        135004 => 1_500,
-        135007 => 2_500,
-        135013 => 2_000,
-        135015 => 2_000,
-        135070 => 2_000,
-        135071 => 3_000,
-        135073 => 1_500,
-        135120 => 2_000,
-        135122 => 2_500,
-        135149 => 1_500,
-        135170 => 3_000,
-        135174 => 2_000,
-        135175 => 2_000,
-        135176 => 1_500,
-        135199 => 2_000,
-        135202 => 3_000,
-        135209 => 2_000,
-        135229 => 3_000,
-        135232 => 1_500,
-        135234 => 2_500,
-        135242 => 2_000,
-        135249 => 2_000,
-        135254 => 3_000,
-        135260 => 2_000,
-        135262 => 1_500,
-        135263 => 2_500,
-        135264 => 2_000,
-        135287 => 2_000,
-        135290 => 1_500,
-        135294 => 2_000,
-        135295 => 2_000,
-        135296 => 2_500,
-        135300 => 1_500,
-        135305 => 2_000,
-        135306 => 2_000,
-        135311 => 1_500,
-        135313 => 3_000,
-        135314 => 3_000,
-        135315 => 2_000,
-        135316 => 1_500,
-        135330 => 2_000,
-        135331 => 2_000,
-        135335 => 3_000,
-        135337 => 2_500,
-        135338 => 2_500,
-        135345 => 3_000,
-        135346 => 1_500,
-        135364 => 3_000,
-        135365 => 1_500,
-        135367 => 3_000,
-        135371 => 1_500,
-        135372 => 2_000,
-        135386 => 3_000,
-        135389 => 2_000,
-        135423 => 2_000,
-        135449 => 2_000,
-        135461 => 3_000,
-        135467 => 2_000,
-        135479 => 3_000,
-        135502 => 3_000,
-        135504 => 3_000,
-        135508 => 3_000,
-        135531 => 2_000,
-        135534 => 2_000,
-        135547 => 2_500,
-        135552 => 2_000,
-        135563 => 3_000,
-        135564 => 2_500,
-        135573 => 3_000,
-        135582 => 1_500,
-        135583 => 2_000,
-        135584 => 2_000,
-        135587 => 2_000,
-        135597 => 2_000,
-        135598 => 2_000,
-        135599 => 3_000,
-        135600 => 1_500,
-        135619 => 1_500,
-        135621 => 1_500,
-        135628 => 2_000,
-        135629 => 1_500,
-        135640 => 3_000,
-        135649 => 1_500,
-        135650 => 2_500,
-        135651 => 1_500,
-        135656 => 3_000,
-        135661 => 3_000,
-        135662 => 2_000,
-        135663 => 1_500,
-        135665 => 2_000,
-        135671 => 2_000,
-        135674 => 1_500,
-        135676 => 2_000,
-        135721 => 2_000,
-        135743 => 3_000,
-        135761 => 3_000,
-        135762 => 2_000,
-        135769 => 3_000,
-        135770 => 3_000,
-        135791 => 2_000,
-        135814 => 1_500,
-        135815 => 2_000,
-        135816 => 2_000,
-        135818 => 2_500,
-        135823 => 1_500,
-        135826 => 2_000,
-        135827 => 2_500,
-        135835 => 2_000,
-        135843 => 2_000,
-        135850 => 3_000,
-        135851 => 3_000,
-        135864 => 2_500,
-        135866 => 2_500,
-        135874 => 3_000,
-        135876 => 1_500,
-        135878 => 3_000,
-        135881 => 3_000,
-        135886 => 2_000,
-        135890 => 2_000,
-        135891 => 2_000,
-        135894 => 2_000,
-        135895 => 2_000,
-        135897 => 2_000,
-        135904 => 3_000,
-        135907 => 2_000,
-        135911 => 2_000,
-        135934 => 2_000,
-        135935 => 3_000,
-        135936 => 2_000,
-        135940 => 1_500,
-        135942 => 3_000,
-        135945 => 1_500,
-        135946 => 2_500,
-        135951 => 1_500,
-        135952 => 3_000,
-        135955 => 2_000,
-        135956 => 3_000,
-        135957 => 2_000,
-        135958 => 2_000,
-        135959 => 1_500,
-        135960 => 1_500,
-        135963 => 2_500,
-        135964 => 1_500,
-        135965 => 3_000,
-        135966 => 2_000,
-        135967 => 2_500,
-        135968 => 2_000,
-        135970 => 2_000,
-        135971 => 1_500,
-        135973 => 1_500,
-        135979 => 2_500,
-        135980 => 2_500,
-        136010 => 2_500,
-        136012 => 2_000,
-        136013 => 2_000,
-        136016 => 1_500,
-        136027 => 3_000,
-        136028 => 1_500,
-        136032 => 2_000,
-        136035 => 2_000,
-        136042 => 2_500,
-        136044 => 2_000,
-        136085 => 2_500,
-        136090 => 2_000,
-        136091 => 3_000,
-        136094 => 1_500,
-        136102 => 3_000,
-        136105 => 3_000,
-        136106 => 3_000,
-        136107 => 1_500,
-        136108 => 2_000,
-        136114 => 1_500,
-        136115 => 3_000,
-        136119 => 3_000,
-        136125 => 1_500,
-        136126 => 3_000,
-        136127 => 2_500,
-        136135 => 2_500,
-        136136 => 1_500,
-        136137 => 2_000,
-        136141 => 2_500,
-        136142 => 3_000,
-        136145 => 2_500,
-        136146 => 2_000,
-        136147 => 2_000,
-        136152 => 2_000,
-        136156 => 2_000,
-        136158 => 1_500,
-        136161 => 2_000,
-        136164 => 2_000,
-        136172 => 2_000,
-        136174 => 3_000,
-        136175 => 2_000,
-        136183 => 1_500,
-        136184 => 2_000,
-        136185 => 3_000,
-        136186 => 2_500,
-        136187 => 3_000,
-        136192 => 1_500,
-        136193 => 2_500,
-        136194 => 2_000,
-        136195 => 2_000,
-        136197 => 2_000,
-        136198 => 2_500,
-        136202 => 2_000,
-        136203 => 2_000,
-        136204 => 2_500,
-        136206 => 1_500,
-        136213 => 2_000,
-        136217 => 3_000,
-        136252 => 3_000,
-        136273 => 2_000,
-        136281 => 2_000,
-        136295 => 2_000,
-        136301 => 2_500,
-        136302 => 1_500,
-        136304 => 2_000,
-        136317 => 1_500,
-        136319 => 2_500,
-        136327 => 1_500,
-        136331 => 2_000,
-        136342 => 1_500,
-        136347 => 1_500,
-        136361 => 2_000,
-        136383 => 2_000,
-        136392 => 2_000,
-        136394 => 2_500,
-        136400 => 2_500,
-        136401 => 3_000,
-        136406 => 2_000,
-        136409 => 3_000,
-        136411 => 3_000,
-        136418 => 1_500,
-        136421 => 2_500,
-        136423 => 2_000,
-        136432 => 2_500,
-        136450 => 2_000,
-        136451 => 2_500,
-        136453 => 3_000,
-        136479 => 2_500,
-        136491 => 2_000,
-        136502 => 3_000,
-        136522 => 1_500,
-        136524 => 2_000,
-        136537 => 2_500,
-        136552 => 2_000,
-        136556 => 2_000,
-        136571 => 2_000,
-        136575 => 2_000,
-        136576 => 1_500,
-        136589 => 2_000,
-        136593 => 2_000,
-        136613 => 3_000,
-        136614 => 2_500,
-        136616 => 2_500,
-        136618 => 2_000,
-        136620 => 2_500,
-        136624 => 2_000,
-        136626 => 3_000,
-        136633 => 2_000,
-        136650 => 2_500,
-        136651 => 2_000,
-        136668 => 3_000,
-        136669 => 1_500,
-        136675 => 2_500,
-        136680 => 2_500,
-        136681 => 3_000,
-        136708 => 2_500,
-        136716 => 2_000,
-        136720 => 1_500,
-        136728 => 2_000,
-        136743 => 3_000,
-        136745 => 3_000,
-        136753 => 2_000,
-        136768 => 1_500,
-        136770 => 3_000,
-        136773 => 2_000,
-        136774 => 2_000,
-        136777 => 2_000,
-        136779 => 2_500,
-        136791 => 2_000,
-        136824 => 2_500,
-        136826 => 3_000,
-        136832 => 1_500,
-        136851 => 2_000,
-        136852 => 2_000,
-        136858 => 1_500,
-        136867 => 1_500,
-        136871 => 2_000,
-        136876 => 1_500,
-        136882 => 2_000,
-        136895 => 3_000,
-        136923 => 1_500,
-        136948 => 1_500,
-        136958 => 2_000,
-        136962 => 2_500,
-        136968 => 3_000,
-        136970 => 2_000,
-        136971 => 2_000,
-        136978 => 2_000,
-        136987 => 3_000,
-        136998 => 2_000,
-        137001 => 2_000,
-        137004 => 1_500,
-        137009 => 1_500,
-        137011 => 2_500,
-        137020 => 2_000,
-        137034 => 2_000,
-        137039 => 3_000,
-        137049 => 2_000,
-        137065 => 3_000,
-        137066 => 3_000,
-        137072 => 1_500,
-        137077 => 1_500,
-        137078 => 1_500,
-        137082 => 2_000,
-        137083 => 2_000,
-        137085 => 2_000,
-        137093 => 1_500,
-        137106 => 2_000,
-        137109 => 2_500,
-        137112 => 2_000,
-        137117 => 2_500,
-        137118 => 2_000,
-        137130 => 2_500,
-        137134 => 3_000,
-        137138 => 1_500,
-        137141 => 2_500,
-        137154 => 1_500,
-        137158 => 2_500,
-        137159 => 2_000,
-        137164 => 2_000,
-        137179 => 2_000,
-        137180 => 2_000,
-        137181 => 2_000,
-        137184 => 2_500,
-        137186 => 1_500,
-        137196 => 2_000,
-        137198 => 1_500,
-        137202 => 2_000,
-        137203 => 2_000,
-        137204 => 1_500,
-        137206 => 2_000,
-        137208 => 2_000,
-        137209 => 2_000,
-        137221 => 1_500,
-        137223 => 2_000,
-        137225 => 2_500,
-        137231 => 1_500,
-        137248 => 2_000,
-        137262 => 2_500,
-        137264 => 2_000,
-        137272 => 2_000,
-        137274 => 3_000,
-        137277 => 2_500,
-        137282 => 2_000,
-        137286 => 2_000,
-        137287 => 1_500,
-        137317 => 2_000,
-        137321 => 2_000,
-        137322 => 3_000,
-        137328 => 2_000,
-        137346 => 2_000,
-        137354 => 1_500,
-        137361 => 2_000,
-        137363 => 2_000,
-        137367 => 1_500,
-        137377 => 2_000,
-        137379 => 1_500,
-        137383 => 2_000,
-        137384 => 1_500,
-        137388 => 1_500,
-        137390 => 3_000,
-        137394 => 2_000,
-        137412 => 2_000,
-        137424 => 2_000,
-        137431 => 2_000,
-        137433 => 3_000,
-        137435 => 2_000,
-        137440 => 1_500,
-        137442 => 1_500,
-        137447 => 2_500,
-        137459 => 3_000,
-        137464 => 1_500,
-        137469 => 2_500,
-        137472 => 2_500,
-        137478 => 2_500,
-        137487 => 2_000,
-        137493 => 1_500,
-        137500 => 3_000,
-        137507 => 2_000,
-        137514 => 2_500,
-        137520 => 2_000,
-        137541 => 2_000,
-        137547 => 2_000,
-        137554 => 2_000,
-        137572 => 2_000,
-        137577 => 2_500,
-        137578 => 2_000,
-        137581 => 1_500,
-        137582 => 1_500,
-        137594 => 3_000,
-        137596 => 2_000,
-        137600 => 2_500,
-        137605 => 1_500,
-        137606 => 2_500,
-        137618 => 1_500,
-        137621 => 2_000,
-        137630 => 1_500,
-        137636 => 2_000,
-        137638 => 2_500,
-        137645 => 2_000,
-        137673 => 2_500,
-        137674 => 2_500,
-        137675 => 3_000,
-        137677 => 2_000,
-        137679 => 2_000,
-        137684 => 2_500,
-        137687 => 2_000,
-        137692 => 1_500,
-        137694 => 1_500,
-        137696 => 2_000,
-        137701 => 3_000,
-        137702 => 2_000,
-        137705 => 2_500,
-        137706 => 3_000,
-        137707 => 2_000,
-        137708 => 1_500,
-        137730 => 2_000,
-        137733 => 1_500,
-        137741 => 2_000,
-        137745 => 1_500,
-        137751 => 3_000,
-        137754 => 1_500,
-        137763 => 2_000,
-        137766 => 1_500,
-        137772 => 2_000,
-        137773 => 2_000,
-        137789 => 2_000,
-        137798 => 2_500,
-        137801 => 2_500,
-        137812 => 2_000,
-        137815 => 2_500,
-        137823 => 3_000,
-        137830 => 3_000,
-        137832 => 1_500,
-        137838 => 2_500,
-        137839 => 2_000,
-        137842 => 2_000,
-        137844 => 1_500,
-        137857 => 2_500,
-        137858 => 1_500,
-        137867 => 2_500,
-        137869 => 2_000,
-        137880 => 2_000,
-        137883 => 1_500,
-        137896 => 2_000,
-        137900 => 1_500,
-        137902 => 2_500,
-        137903 => 2_500,
-        137905 => 1_500,
-        137907 => 1_500,
-        137911 => 2_000,
-        137912 => 2_000,
-        137915 => 3_000,
-        137934 => 1_500,
-        137937 => 1_500,
-        137949 => 2_000,
-        137954 => 2_000,
-        137966 => 2_000,
-        137973 => 2_500,
-        137974 => 2_000,
-        137981 => 2_000,
-        137998 => 2_000,
-        138003 => 2_000,
-        138006 => 2_500,
-        138023 => 2_000,
-        138039 => 2_000,
-        138044 => 2_000,
-        138046 => 1_500,
-        138053 => 2_000,
-        138059 => 2_000,
-        138067 => 2_000,
-        138075 => 2_500,
-        138076 => 2_000,
-        138082 => 1_500,
-        138087 => 2_500,
-        138089 => 3_000,
-        138093 => 3_000,
-        138094 => 2_000,
-        138097 => 3_000,
-        138098 => 2_000,
-        138099 => 3_000,
-        138102 => 1_500,
-        138103 => 2_500,
-        138107 => 1_500,
-        138116 => 2_500,
-        138120 => 2_000,
-        138122 => 1_500,
-        138124 => 2_000,
-        138130 => 3_000,
-        138137 => 2_000,
-        138156 => 1_500,
-        138158 => 2_000,
-        138162 => 2_000,
-        138167 => 2_000,
-        138177 => 2_000,
-        138183 => 2_500,
-        138186 => 2_000,
-        138190 => 2_000,
-        138196 => 2_000,
-        138202 => 2_000,
-        138204 => 1_500,
-        138208 => 1_500,
-        138219 => 1_500,
-        138221 => 1_500,
-        138222 => 2_000,
-        138228 => 2_000,
-        138233 => 2_000,
-        138236 => 1_500,
-        138245 => 1_500,
-        138246 => 2_000,
-        138248 => 2_000,
-        138250 => 2_000,
-        138251 => 2_000,
-        138253 => 2_000,
-        138254 => 2_000,
-        138260 => 3_000,
-        138262 => 2_000,
-        138266 => 1_500,
-        138271 => 3_000,
-        138273 => 2_500,
-        138274 => 3_000,
-        138277 => 3_000,
-        138281 => 2_000,
-        138304 => 2_000,
-        138312 => 1_500,
-        138314 => 2_500,
-        138327 => 1_500,
-        138329 => 2_000,
-        138336 => 2_000,
-        138341 => 2_000,
-        138355 => 2_000,
-        138360 => 1_500,
-        138367 => 1_500,
-        138368 => 1_500,
-        138373 => 2_500,
-        138374 => 3_000,
-        138375 => 2_500,
-        138379 => 3_000,
-        138380 => 2_500,
-        138411 => 2_500,
-        138414 => 2_000,
-        138435 => 1_500,
-        138437 => 1_500,
-        138439 => 2_000,
-        138448 => 2_000,
-        138449 => 1_500,
-        138450 => 2_000,
-        138454 => 2_000,
-        138455 => 3_000,
-        138463 => 2_000,
-        138472 => 2_000,
-        138474 => 2_000,
-        138475 => 2_000,
-        138485 => 1_500,
-        138495 => 2_000,
-        138496 => 1_500,
-        138499 => 1_500,
-        138506 => 2_500,
-        138511 => 2_000,
-        138518 => 1_500,
-        138526 => 3_000,
-        138530 => 3_000,
-        138532 => 3_000,
-        138545 => 1_500,
-        138547 => 2_000,
-        138549 => 3_000,
-        138565 => 2_000,
-        138567 => 2_000,
-        138585 => 1_500,
-        138586 => 1_500,
-        138593 => 2_500,
-        138606 => 3_000,
-        138620 => 1_500,
-        138622 => 2_000,
-        138634 => 3_000,
-        138644 => 1_500,
-        138652 => 2_000,
-        138661 => 2_500,
-        138681 => 2_000,
-        138682 => 2_000,
-        138685 => 2_000,
-        138686 => 1_500,
-        138695 => 2_000,
-        138696 => 2_500,
-        138697 => 3_000,
-        138706 => 2_000,
-        138707 => 2_000,
-        138711 => 3_000,
-        138716 => 2_500,
-        138717 => 2_500,
-        138718 => 3_000,
-        138736 => 1_500,
-        138747 => 1_500,
-        138754 => 2_000,
-        138756 => 3_000,
-        138758 => 2_000,
-        138766 => 2_000,
-        138775 => 2_000,
-        138783 => 2_500,
-        138787 => 3_000,
-        138810 => 2_000,
-        138832 => 2_000,
-        138836 => 2_500,
-        138837 => 2_500,
-        138853 => 2_500,
-        138859 => 1_500,
-        138863 => 1_500,
-        138865 => 2_000,
-        138885 => 1_500,
-        138895 => 2_500,
-        138918 => 2_000,
-        138920 => 1_500,
-        138923 => 2_500,
-        138924 => 1_500,
-        138925 => 2_000,
-        138935 => 1_500,
-        138937 => 2_000,
-        138950 => 2_000,
-        138960 => 1_500,
-        138961 => 1_500,
-        138985 => 1_500,
-        139031 => 2_000,
-        139037 => 2_000,
-        139043 => 2_500,
-        139047 => 2_000,
-        139048 => 2_000,
-        139049 => 2_000,
-        139058 => 2_000,
-        139059 => 2_000,
-        139060 => 2_000,
-        139062 => 2_500,
-        139063 => 3_000,
-        139067 => 3_000,
-        139068 => 2_500,
-        139076 => 2_500,
-        139082 => 3_000,
-        139095 => 1_500,
-        139099 => 3_000,
-        139110 => 2_000,
-        139114 => 3_000,
-        139118 => 1_500,
-        139130 => 1_500,
-        139138 => 2_000,
-        139148 => 2_500,
-        139155 => 2_000,
-        139160 => 2_000,
-        139184 => 2_000,
-        139196 => 2_000,
-        139197 => 2_000,
-        139210 => 1_500,
-        139211 => 1_500,
-        139258 => 2_000,
-        139272 => 2_500,
-        139276 => 1_500,
-        139277 => 1_500,
-        139288 => 2_500,
-        139293 => 1_500,
-        139294 => 3_000,
-        139307 => 1_500,
-        139312 => 2_000,
-        139316 => 3_000,
-        139320 => 3_000,
-        139322 => 1_500,
-        139334 => 2_000,
-        139362 => 1_500,
-        139364 => 2_000,
-        139365 => 2_000,
-        139368 => 2_000,
-        139369 => 1_500,
-        139374 => 3_000,
-        139376 => 2_000,
-        139395 => 2_000,
-        139401 => 3_000,
-        139403 => 3_000,
-        139405 => 3_000,
-        139411 => 3_000,
-        139413 => 2_000,
-        139416 => 1_500,
-        139418 => 2_000,
-        139434 => 1_500,
-        139456 => 3_000,
-        139469 => 2_000,
-        139474 => 3_000,
-        139482 => 3_000,
-        139483 => 2_000,
-        139498 => 2_000,
-        139500 => 2_000,
-        139501 => 2_000,
-        139509 => 2_000,
-        139517 => 2_000,
-        139521 => 3_000,
-        139526 => 2_000,
-        139536 => 2_000,
-        139540 => 2_000,
-        139560 => 3_000,
-        139572 => 2_000,
-        139578 => 1_500,
-        139587 => 2_000,
-        139588 => 3_000,
-        139589 => 3_000,
-        139599 => 1_500,
-        139600 => 2_000,
-        139607 => 2_000,
-        139614 => 3_000,
-        139616 => 1_500,
-        139623 => 2_500,
-        139625 => 2_500,
-        139628 => 2_000,
-        139629 => 2_000,
-        139646 => 3_000,
-        139648 => 2_000,
-        139649 => 2_000,
-        139659 => 2_000,
-        139660 => 2_000,
-        139665 => 3_000,
-        139671 => 2_000,
-        139673 => 2_000,
-        139675 => 3_000,
-        139703 => 2_000,
-        139714 => 3_000,
-        139718 => 2_000,
-        139722 => 2_000,
-        139728 => 2_000,
-        139731 => 2_000,
-        139732 => 2_000,
-        139735 => 2_500,
-        139740 => 3_000,
-        139741 => 3_000,
-        139746 => 1_500,
-        139758 => 3_000,
-        139759 => 3_000,
-        139760 => 3_000,
-        139761 => 3_000,
-        139765 => 3_000,
-        139766 => 2_000,
-        139788 => 2_000,
-        139789 => 3_000,
-        139791 => 1_500,
-        139815 => 2_000,
-        139819 => 2_500,
-        139821 => 2_500,
-        139823 => 3_000,
-        139839 => 2_000,
-        139841 => 2_000,
-        139842 => 1_500,
-        139852 => 2_500,
-        139856 => 2_000,
-        139867 => 2_000,
-        139878 => 2_000,
-        139888 => 2_000,
-        139891 => 1_500,
-        139903 => 2_000,
-        139918 => 3_000,
-        139924 => 2_000,
-        139953 => 2_000,
-        139956 => 2_500,
-        139957 => 2_500,
-        139961 => 2_500,
-        139974 => 2_000,
-        139976 => 3_000,
-        139977 => 2_500,
-        139978 => 3_000,
-        139992 => 2_500,
-        139994 => 2_000,
-        139995 => 3_000,
-        140001 => 3_000,
-        140002 => 2_000,
-        140014 => 2_000,
-        140021 => 3_000,
-        140032 => 2_500,
-        140047 => 2_000,
-        140055 => 2_000,
-        140067 => 2_000,
-        140077 => 3_000,
-        140079 => 2_000,
-        140091 => 2_000,
-        140093 => 2_000,
-        140104 => 2_000,
-        140121 => 3_000,
-        140126 => 2_500,
-        140138 => 2_000,
-        140143 => 3_000,
-        140149 => 2_500,
-        140160 => 2_000,
-        140197 => 2_000,
-        140201 => 2_000,
-        140204 => 2_500,
-        140209 => 2_000,
-        140210 => 2_000,
-        140212 => 2_000,
-        140217 => 1_500,
-        140221 => 2_000,
-        140248 => 2_500,
-        140252 => 2_000,
-        140261 => 2_000,
-        140277 => 2_500,
-        140286 => 3_000,
-        140304 => 2_500,
-        140308 => 1_500,
-        140325 => 3_000,
-        140326 => 1_500,
-        140360 => 2_000,
-        140364 => 2_500,
-        140366 => 2_500,
-        140369 => 3_000,
-        140373 => 2_000,
-        140388 => 3_000,
-        140394 => 2_000,
-        140397 => 2_000,
-        140404 => 2_000,
-        140412 => 3_000,
-        140415 => 1_500,
-        140416 => 2_000,
-        140429 => 3_000,
-        140436 => 2_000,
-        140457 => 2_000,
-        140459 => 2_000,
-        140475 => 1_500,
-        140500 => 3_000,
-        140515 => 2_000,
-        140531 => 1_500,
-        140549 => 3_000,
-        140553 => 2_500,
-        140557 => 2_000,
-        140565 => 3_000,
-        140569 => 2_000,
-        140578 => 3_000,
-        140586 => 2_000,
-        140600 => 1_500,
-        140605 => 2_500,
-        140627 => 2_000,
-        140631 => 2_500,
-        140633 => 2_500,
-        140649 => 2_000,
-        140654 => 2_000,
-        140660 => 2_000,
-        140669 => 2_000,
-        140675 => 2_000,
-        140677 => 2_000,
-        140679 => 2_500,
-        140693 => 2_500,
-        140697 => 2_000,
-        140701 => 2_000,
-        140703 => 3_000,
-        140731 => 2_000,
-        140732 => 2_000,
-        140737 => 2_500,
-        140751 => 2_000,
-        140753 => 3_000,
-        140759 => 2_500,
-        140795 => 2_000,
-        140796 => 2_000,
-        140806 => 2_000,
-        140807 => 2_000,
-        140813 => 1_500,
-        140815 => 2_500,
-        140821 => 2_000,
-        140836 => 2_000,
-        140845 => 2_000,
-        140854 => 3_000,
-        140861 => 2_000,
-        140862 => 1_500,
-        140863 => 1_500,
-        140864 => 3_000,
-        140865 => 2_000,
-        140867 => 2_000,
-        140879 => 1_500,
-        140881 => 2_000,
-        140882 => 3_000,
-        140884 => 2_000,
-        140898 => 2_000,
-        140904 => 2_000,
-        140907 => 2_000,
-        140935 => 2_000,
-        140950 => 2_000,
-        140952 => 2_000,
-        140957 => 2_000,
-        140958 => 3_000,
-        140959 => 2_000,
-        140961 => 2_000,
-        140962 => 1_500,
-        140964 => 2_000,
-        140965 => 2_000,
-        140966 => 2_000,
-        140977 => 2_000,
-        140984 => 2_500,
-        140985 => 2_000,
-        140987 => 1_500,
-        140992 => 2_500,
-        140997 => 2_000,
-        141002 => 2_500,
-        141003 => 2_000,
-        141004 => 1_500,
-        141005 => 2_000,
-        141006 => 1_500,
-        141010 => 2_000,
-        141016 => 2_000,
-        141019 => 1_500,
-        141028 => 1_500,
-        141033 => 3_000,
-        141034 => 2_000,
-        141037 => 2_000,
-        141041 => 1_500,
-        141071 => 1_500,
-        141082 => 2_000,
-        141086 => 2_000,
-        141101 => 2_000,
-        141103 => 2_500,
-        141109 => 3_000,
-        141116 => 2_000,
-        141123 => 2_000,
-        141130 => 2_000,
-        141132 => 3_000,
-        141133 => 2_000,
-        141139 => 1_500,
-        141140 => 3_000,
-        141141 => 3_000,
-        141147 => 1_500,
-        141163 => 1_500,
-        141164 => 2_000,
-        141166 => 2_000,
-        141171 => 2_000,
-        141175 => 2_500,
-        141178 => 1_500,
-        141179 => 2_000,
-        141187 => 2_000,
-        141196 => 3_000,
-        141200 => 1_500,
-        141210 => 2_000,
-        141214 => 1_500,
-        141217 => 1_500,
-        141236 => 2_500,
-        141245 => 1_500,
-        141248 => 2_000,
-        141251 => 1_500,
-        141252 => 2_000,
-        141253 => 3_000,
-        141254 => 3_000,
-        141255 => 2_000,
-        141257 => 2_500,
-        141259 => 2_500,
-        141264 => 3_000,
-        141268 => 2_000,
-        141269 => 2_000,
-        141300 => 2_000,
-        141318 => 2_000,
-        141321 => 1_500,
-        141325 => 2_500,
-        141331 => 2_500,
-        141338 => 2_500,
-        141363 => 3_000,
-        141376 => 3_000,
-        141377 => 2_500,
-        141378 => 2_000,
-        141385 => 2_000,
-        141391 => 2_500,
-        141395 => 2_000,
-        141407 => 3_000,
-        141448 => 2_000,
-        141449 => 1_500,
-        141470 => 2_500,
-        141475 => 3_000,
-        141485 => 2_000,
-        141487 => 2_000,
-        141495 => 1_500,
-        141499 => 1_500,
-        141512 => 2_000,
-        141514 => 1_500,
-        141559 => 1_500,
-        141563 => 2_000,
-        141564 => 2_000,
-        141570 => 2_000,
-        141581 => 2_000,
-        141585 => 2_000,
-        141605 => 2_000,
-        141606 => 2_000,
-        141617 => 2_000,
-        141647 => 1_500,
-        141652 => 1_500,
-        141665 => 2_000,
-        141668 => 2_000,
-        141683 => 1_500,
-        141686 => 3_000,
-        141693 => 3_000,
-        141694 => 3_000,
-        141698 => 2_500,
-        141704 => 2_000,
-        141705 => 1_500,
-        141712 => 2_000,
-        141726 => 2_000,
-        141727 => 2_000,
-        141730 => 2_500,
-        141735 => 2_000,
-        141741 => 2_000,
-        141749 => 2_000,
-        141752 => 2_000,
-        141765 => 2_000,
-        141791 => 3_000,
-        141802 => 2_000,
-        141805 => 3_000,
-        141813 => 3_000,
-        141835 => 2_000,
-        141843 => 1_500,
-        141853 => 2_500,
-        141865 => 2_000,
-        141874 => 1_500,
-        141875 => 2_500,
-        141883 => 3_000,
-        141884 => 3_000,
-        141935 => 2_000,
-        141939 => 2_000,
-        141941 => 3_000,
-        141948 => 3_000,
-        141957 => 2_000,
-        141969 => 1_500,
-        141986 => 3_000,
-        141989 => 2_000,
-        141992 => 2_500,
-        141993 => 2_000,
-        142008 => 2_000,
-        142031 => 2_500,
-        142035 => 2_000,
-        142040 => 2_000,
-        142042 => 2_500,
-        142050 => 2_000,
-        142051 => 1_500,
-        142054 => 2_000,
-        142056 => 2_500,
-        142058 => 3_000,
-        142059 => 3_000,
-        142071 => 2_000,
-        142073 => 2_500,
-        142075 => 1_500,
-        142079 => 2_500,
-        142085 => 2_000,
-        142086 => 2_000,
-        142094 => 1_500,
-        142107 => 2_000,
-        142118 => 2_500,
-        142125 => 1_500,
-        142126 => 2_000,
-        142127 => 1_500,
-        142130 => 2_000,
-        142134 => 2_000,
-        142135 => 2_000,
-        142136 => 3_000,
-        142143 => 3_000,
-        142146 => 3_000,
-        142151 => 3_000,
-        142163 => 3_000,
-        142170 => 2_000,
-        142174 => 2_000,
-        142178 => 1_500,
-        142186 => 3_000,
-        142219 => 2_000,
-        142232 => 2_500,
-        142257 => 2_000,
-        142258 => 2_000,
-        142260 => 2_000,
-        142277 => 1_500,
-        142281 => 2_000,
-        142285 => 2_000,
-        142288 => 2_000,
-        142296 => 2_000,
-        142302 => 2_000,
-        142308 => 3_000,
-        142309 => 3_000,
-        142339 => 2_500,
-        142340 => 2_000,
-        142378 => 3_000,
-        142382 => 3_000,
-        142383 => 2_000,
-        142387 => 2_500,
-        142388 => 2_000,
-        142391 => 2_000,
-        142392 => 2_500,
-        142396 => 2_000,
-        142405 => 3_000,
-        142406 => 2_500,
-        142420 => 3_000,
-        142469 => 3_000,
-        142485 => 1_500,
-        142487 => 3_000,
-        142502 => 3_000,
-        142510 => 3_000,
-        142540 => 1_500,
-        142543 => 1_500,
-        142555 => 2_000,
-        142557 => 3_000,
-        142569 => 2_000,
-        142585 => 3_000,
-        142590 => 2_000,
-        142594 => 3_000,
-        142601 => 2_000,
-        142604 => 2_500,
-        142605 => 1_500,
-        142607 => 2_000,
-        142612 => 2_000,
-        142630 => 3_000,
-        142644 => 2_000,
-        142666 => 3_000,
-        142667 => 3_000,
-        142681 => 3_000,
-        142687 => 1_500,
-        142695 => 2_000,
-        142698 => 1_500,
-        142719 => 3_000,
-        142722 => 1_500,
-        142725 => 2_000,
-        142740 => 1_500,
-        142741 => 2_000,
-        142744 => 3_000,
-        142747 => 1_500,
-        142758 => 3_000,
-        142759 => 2_500,
-        142761 => 3_000,
-        142762 => 3_000,
-        142766 => 3_000,
-        142768 => 3_000,
-        142780 => 2_000,
-        142781 => 1_500,
-        142783 => 2_000,
-        142793 => 3_000,
-        142795 => 3_000,
-        142797 => 2_500,
-        142798 => 3_000,
-        142801 => 3_000,
-        142808 => 3_000,
-        142820 => 3_000,
-        142825 => 3_000,
-        142835 => 3_000,
-        142853 => 2_000,
-        142879 => 1_500,
-        142881 => 1_500,
-        142882 => 3_000,
-        142883 => 3_000,
-        142884 => 2_000,
-        142887 => 1_500,
-        142890 => 1_500,
-        142893 => 3_000,
-        142894 => 2_500,
-        142899 => 2_000,
-        142900 => 1_500,
-        142902 => 2_500,
-        142903 => 2_000,
-        142905 => 2_000,
-        142906 => 2_000,
-        142908 => 1_500,
-        142926 => 2_000,
-        142928 => 2_000,
-        142932 => 3_000,
-        142938 => 2_000,
-        142940 => 2_000,
-        142960 => 2_500,
-        143014 => 2_500,
-        143035 => 2_000,
-        143059 => 3_000,
-        143062 => 3_000,
-        143082 => 1_500,
-        143111 => 2_500,
-        143112 => 2_500,
-        143114 => 2_500,
-        143115 => 2_000,
-        143129 => 2_000,
-        143130 => 2_000,
-        143137 => 2_000,
-        143141 => 2_000,
-        143144 => 2_000,
-        143146 => 2_000,
-        143166 => 3_000,
-        143167 => 3_000,
-        143173 => 3_000,
-        143175 => 2_000,
-        143197 => 1_500,
-        143217 => 2_000,
-        143222 => 2_000,
-        143227 => 3_000,
-        143247 => 2_000,
-        143260 => 2_500,
-        143274 => 1_500,
-        143275 => 2_000,
-        143288 => 2_000,
-        143317 => 2_000,
-        143319 => 2_500,
-        143327 => 1_500,
-        143346 => 1_500,
-        143362 => 2_000,
-        143379 => 2_000,
-        143385 => 1_500,
-        143413 => 2_000,
-        143420 => 1_500,
-        143424 => 1_500,
-        143427 => 2_000,
-        143435 => 2_000,
-        143438 => 2_000,
-        143472 => 3_000,
-        143490 => 3_000,
-        143519 => 3_000,
-        143520 => 2_000,
-        143533 => 2_500,
-        143545 => 3_000,
-        143552 => 2_000,
-        143554 => 3_000,
-        143574 => 2_000,
-        143582 => 2_000,
-        143589 => 1_500,
-        143590 => 2_000,
-        143591 => 2_000,
-        143595 => 2_000,
-        143629 => 2_000,
-        143630 => 2_000,
-        143639 => 1_500,
-        143641 => 2_000,
-        143649 => 2_000,
-        143659 => 2_000,
-        143686 => 1_500,
-        143702 => 2_000,
-        143714 => 2_000,
-        143729 => 2_500,
-        143731 => 1_500,
-        143734 => 2_500,
-        143745 => 2_000,
-        143746 => 3_000,
-        143750 => 3_000,
-        143756 => 2_000,
-        143772 => 3_000,
-        143773 => 2_000,
-        143794 => 1_500,
-        143818 => 2_000,
-        143828 => 3_000,
-        143844 => 1_500,
-        143845 => 3_000,
-        143846 => 2_000,
-        143848 => 2_000,
-        143853 => 3_000,
-        143857 => 1_500,
-        143865 => 2_000,
-        143879 => 1_500,
-        143895 => 2_000,
-        143907 => 1_500,
-        143916 => 2_000,
-        143918 => 2_000,
-        143927 => 2_000,
-        143934 => 2_000,
-        143938 => 2_500,
-        143946 => 1_500,
-        143963 => 1_500,
-        144014 => 1_500,
-        144015 => 1_500,
-        144022 => 3_000,
-        144023 => 3_000,
-        144031 => 2_000,
-        144055 => 3_000,
-        144066 => 3_000,
-        144069 => 2_500,
-        144084 => 2_000,
-        144134 => 2_000,
-        144182 => 2_000,
-        144196 => 3_000,
-        144199 => 2_000,
-        144200 => 2_000,
-        144204 => 2_500,
-        144205 => 2_000,
-        144209 => 2_000,
-        144234 => 2_000,
-        144240 => 2_000,
-        144271 => 2_000,
-        144278 => 2_000,
-        144279 => 1_500,
-        144284 => 2_000,
-        144285 => 2_000,
-        144286 => 2_000,
-        144291 => 2_000,
-        144306 => 2_000,
-        144307 => 2_000,
-        144309 => 1_500,
-        144311 => 2_000,
-        144312 => 2_000,
-        144320 => 3_000,
-        144321 => 3_000,
-        144336 => 2_000,
-        144342 => 2_000,
-        144346 => 3_000,
-        144352 => 2_000,
-        144356 => 2_000,
-        144358 => 2_000,
-        144359 => 2_000,
-        144364 => 2_500,
-        144367 => 1_500,
-        144391 => 2_500,
-        144392 => 1_500,
-        144405 => 2_000,
-        144407 => 2_000,
-        144422 => 2_000,
-        144464 => 2_000,
-        144479 => 2_000,
-        144486 => 3_000,
-        144487 => 3_000,
-        144488 => 2_000,
-        144489 => 2_000,
-        144493 => 3_000,
-        144494 => 3_000,
-        144502 => 1_500,
-        144505 => 2_000,
-        144508 => 3_000,
-        144509 => 2_000,
-        144519 => 2_000,
-        144533 => 2_500,
-        144588 => 2_500,
-        144601 => 1_500,
-        144606 => 1_500,
-        144610 => 1_500,
-        144617 => 3_000,
-        144618 => 2_000,
-        144627 => 2_500,
-        144629 => 1_500,
-        144633 => 2_000,
-        144643 => 3_000,
-        144646 => 2_000,
-        144652 => 1_500,
-        144655 => 2_000,
-        144657 => 3_000,
-        144661 => 2_000,
-        144684 => 3_000,
-        144687 => 2_000,
-        144691 => 2_000,
-        144692 => 2_000,
-        144693 => 2_000,
-        144700 => 2_000,
-        144719 => 2_000,
-        144721 => 2_000,
-        144739 => 2_000,
-        144743 => 3_000,
-        144744 => 3_000,
-        144745 => 2_500,
-        144748 => 1_500,
-        144755 => 3_000,
-        144756 => 2_000,
-        144759 => 1_500,
-        144760 => 2_500,
-        144763 => 3_000,
-        144764 => 1_500,
-        144769 => 3_000,
-        144773 => 2_500,
-        144784 => 2_000,
-        144799 => 1_500,
-        144809 => 3_000,
-        144852 => 2_000,
-        144875 => 2_000,
-        144879 => 3_000,
-        144893 => 2_000,
-        144897 => 2_000,
-        144900 => 2_000,
-        144937 => 3_000,
-        144946 => 2_000,
-        144952 => 3_000,
-        144956 => 1_500,
-        144962 => 2_000,
-        144968 => 2_500,
-        144977 => 1_500,
-        144988 => 1_500,
-        144990 => 3_000,
-        144991 => 2_000,
-        145008 => 2_000,
-        145034 => 2_500,
-        145051 => 3_000,
-        145052 => 2_000,
-        145055 => 2_000,
-        145058 => 2_000,
-        145061 => 1_500,
-        145089 => 2_000,
-        145106 => 1_500,
-        145108 => 2_500,
-        145109 => 2_500,
-        145113 => 2_000,
-        145114 => 1_500,
-        145119 => 2_500,
-        145120 => 2_000,
-        145121 => 2_000,
-        145124 => 2_000,
-        145125 => 1_500,
-        145126 => 2_000,
-        145132 => 3_000,
-        145133 => 2_000,
-        145134 => 2_000,
-        145135 => 2_000,
-        145139 => 2_000,
-        145140 => 2_000,
-        145146 => 2_500,
-        145155 => 2_500,
-        145173 => 3_000,
-        145179 => 2_500,
-        145188 => 2_000,
-        145196 => 2_000,
-        145198 => 2_000,
-        145200 => 1_500,
-        145211 => 2_000,
-        145224 => 1_500,
-        145232 => 2_000,
-        145253 => 3_000,
-        145254 => 2_000,
-        145271 => 2_500,
-        145297 => 1_500,
-        145313 => 2_000,
-        145314 => 3_000,
-        145315 => 1_500,
-        145327 => 2_500,
-        145361 => 2_000,
-        145372 => 3_000,
-        145383 => 2_000,
-        145384 => 1_500,
-        145410 => 2_000,
-        145420 => 2_000,
-        145430 => 2_000,
-        145433 => 3_000,
-        145434 => 3_000,
-        145477 => 3_000,
-        145480 => 1_500,
-        145486 => 1_500,
-        145490 => 2_000,
-        145494 => 2_000,
-        145501 => 2_000,
-        145521 => 2_000,
-        145528 => 2_000,
-        145529 => 3_000,
-        145564 => 2_500,
-        145577 => 2_500,
-        145579 => 2_500,
-        145580 => 1_500,
-        145581 => 3_000,
-        145591 => 2_000,
-        145593 => 2_500,
-        145677 => 2_500,
-        145704 => 2_000,
-        145708 => 1_500,
-        145709 => 2_000,
-        145723 => 3_000,
-        145732 => 3_000,
-        145736 => 1_500,
-        145741 => 1_500,
-        145774 => 2_500,
-        145843 => 2_000,
-        145845 => 3_000,
-        145848 => 2_000,
-        145850 => 3_000,
-        145852 => 3_000,
-        145857 => 2_000,
-        145860 => 1_500,
-        145861 => 2_000,
-        145863 => 2_500,
-        145864 => 2_000,
-        145866 => 1_500,
-        145872 => 2_500,
-        145873 => 3_000,
-        145874 => 3_000,
-        145878 => 1_500,
-        145894 => 2_000,
-        145898 => 1_500,
-        145900 => 1_500,
-        145906 => 3_000,
-        145917 => 2_000,
-        145918 => 2_000,
-        145921 => 2_000,
-        145922 => 3_000,
-        145923 => 1_500,
-        145927 => 3_000,
-        145937 => 2_000,
-        145938 => 2_500,
-        145943 => 1_500,
-        145944 => 1_500,
-        145952 => 3_000,
-        145954 => 3_000,
-        145957 => 2_000,
-        145966 => 2_000,
-        145987 => 2_000,
-        146013 => 2_000,
-        146014 => 2_000,
-        146052 => 3_000,
-        146053 => 3_000,
-        146054 => 2_500,
-        146055 => 3_000,
-        146056 => 3_000,
-        146057 => 3_000,
-        146058 => 3_000,
-        146070 => 1_500,
-        146087 => 2_000,
-        146099 => 2_500,
-        146121 => 1_500,
-        146124 => 2_000,
-        146128 => 2_000,
-        146190 => 2_000,
-        146198 => 2_500,
-        146204 => 2_000,
-        146209 => 2_000,
-        146212 => 2_000,
-        146217 => 2_000,
-        146220 => 2_000,
-        146237 => 2_000,
-        146239 => 3_000,
-        146247 => 2_000,
-        146255 => 1_500,
-        146272 => 2_000,
-        146274 => 2_000,
-        146284 => 1_500,
-        146303 => 3_000,
-        146305 => 1_500,
-        146309 => 1_500,
-        146315 => 3_000,
-        146316 => 3_000,
-        146318 => 2_000,
-        146325 => 2_500,
-        146327 => 2_500,
-        146363 => 3_000,
-        146365 => 2_500,
-        146367 => 2_500,
-        146369 => 2_500,
-        146370 => 3_000,
-        146374 => 2_500,
-        146375 => 2_000,
-        146378 => 2_000,
-        146379 => 2_000,
-        146383 => 3_000,
-        146404 => 2_000,
-        146408 => 2_000,
-        146410 => 3_000,
-        146418 => 2_000,
-        146422 => 3_000,
-        146431 => 3_000,
-        146436 => 2_500,
-        146439 => 2_000,
-        146440 => 1_500,
-        146451 => 2_000,
-        146493 => 1_500,
-        146497 => 3_000,
-        146501 => 2_000,
-        146503 => 3_000,
-        146518 => 1_500,
-        146527 => 2_000,
-        146529 => 2_000,
-        146539 => 3_000,
-        146560 => 3_000,
-        146562 => 2_000,
-        146564 => 3_000,
-        146567 => 1_500,
-        146601 => 2_000,
-        146624 => 1_500,
-        146634 => 1_500,
-        146635 => 1_500,
-        146639 => 3_000,
-        146640 => 3_000,
-        146648 => 1_500,
-        146680 => 2_000,
-        146683 => 1_500,
-        146698 => 2_500,
-        146699 => 1_500,
-        146727 => 2_000,
-        146731 => 2_000,
-        146732 => 2_000,
-        146734 => 2_000,
-        146735 => 2_000,
-        146742 => 2_000,
-        146752 => 1_500,
-        146765 => 2_000,
-        146769 => 2_500,
-        146793 => 3_000,
-        146797 => 2_000,
-        146800 => 1_500,
-        146819 => 2_500,
-        146820 => 1_500,
-        146825 => 2_500,
-        146851 => 3_000,
-        146858 => 2_000,
-        146885 => 2_000,
-        146891 => 3_000,
-        146899 => 2_000,
-        146901 => 1_500,
-        146909 => 2_500,
-        146914 => 1_500,
-        146936 => 1_500,
-        146953 => 2_000,
-        146958 => 3_000,
-        146969 => 2_000,
-        146970 => 3_000,
-        147016 => 2_000,
-        147018 => 2_000,
-        147031 => 1_500,
-        147061 => 1_500,
-        147064 => 2_000,
-        147067 => 2_500,
-        147068 => 1_500,
-        147071 => 2_000,
-        147072 => 1_500,
-        147073 => 2_000,
-        147077 => 2_500,
-        147080 => 1_500,
-        147087 => 3_000,
-        147100 => 2_500,
-        147115 => 1_500,
-        147116 => 2_000,
-        147122 => 3_000,
-        147124 => 3_000,
-        147126 => 2_500,
-        147132 => 3_000,
-        147136 => 2_000,
-        147137 => 3_000,
-        147143 => 3_000,
-        147157 => 2_000,
-        147158 => 1_500,
-        147162 => 3_000,
-        147177 => 3_000,
-        147178 => 2_000,
-        147182 => 2_500,
-        147185 => 2_000,
-        147188 => 1_500,
-        147189 => 2_000,
-        147195 => 3_000,
-        147196 => 3_000,
-        147197 => 3_000,
-        147201 => 2_000,
-        147204 => 2_000,
-        147214 => 2_500,
-        147219 => 2_000,
-        147260 => 1_500,
-        147268 => 2_500,
-        147291 => 2_500,
-        147300 => 1_500,
-        147345 => 2_000,
-        147346 => 2_000,
-        147347 => 3_000,
-        147350 => 2_000,
-        147351 => 2_500,
-        147362 => 2_000,
-        147369 => 2_500,
-        147400 => 2_500,
-        147401 => 1_500,
-        147414 => 2_500,
-        147439 => 2_000,
-        147440 => 2_000,
-        147442 => 2_500,
-        147443 => 2_000,
-        147451 => 2_000,
-        147456 => 3_000,
-        147467 => 1_500,
-        147491 => 3_000,
-        147505 => 2_000,
-        147515 => 3_000,
-        147524 => 1_500,
-        147531 => 2_000,
-        147532 => 2_000,
-        147535 => 1_500,
-        147536 => 1_500,
-        147538 => 1_500,
-        147540 => 1_500,
-        147541 => 2_000,
-        147543 => 2_000,
-        147544 => 2_000,
-        147565 => 1_500,
-        147577 => 2_000,
-        147582 => 2_500,
-        147594 => 2_500,
-        147627 => 3_000,
-        147653 => 2_000,
-        147654 => 2_000,
-        147661 => 2_000,
-        147670 => 2_000,
-        147675 => 2_000,
-        147676 => 2_000,
-        147679 => 2_000,
-        147683 => 2_000,
-        147685 => 3_000,
-        147707 => 2_000,
-        147711 => 2_500,
-        147719 => 2_000,
-        147720 => 3_000,
-        147728 => 2_000,
-        147757 => 2_000,
-        147793 => 2_000,
-        147796 => 3_000,
-        147801 => 1_500,
-        147819 => 2_500,
-        147821 => 1_500,
-        147829 => 3_000,
-        147832 => 2_000,
-        147838 => 2_000,
-        147840 => 2_000,
-        147841 => 3_000,
-        147845 => 3_000,
-        147848 => 3_000,
-        147849 => 3_000,
-        147853 => 1_500,
-        147854 => 2_000,
-        147855 => 3_000,
-        147863 => 1_500,
-        147864 => 3_000,
-        147874 => 2_000,
-        147875 => 3_000,
-        147881 => 3_000,
-        147883 => 2_500,
-        147888 => 3_000,
-        147889 => 2_000,
-        147892 => 3_000,
-        147894 => 2_500,
-        147897 => 1_500,
-        147905 => 3_000,
-        147942 => 1_500,
-        147944 => 3_000,
-        147967 => 3_000,
-        148003 => 2_000,
-        148008 => 2_000,
-        148029 => 2_000,
-        148035 => 3_000,
-        148036 => 2_000,
-        148055 => 2_500,
-        148109 => 2_000,
-        148124 => 2_000,
-        148129 => 2_000,
-        148135 => 2_000,
-        148144 => 2_000,
-        148145 => 2_000,
-        148163 => 2_000,
-        148165 => 2_000,
-        148167 => 2_000,
-        148171 => 2_000,
-        148179 => 2_500,
-        148185 => 2_000,
-        148186 => 2_000,
-        148187 => 2_000,
-        148220 => 3_000,
-        148224 => 2_000,
-        148225 => 2_000,
-        148226 => 3_000,
-        148235 => 2_000,
-        148236 => 3_000,
-        148250 => 2_000,
-        148267 => 2_500,
-        148294 => 2_500,
-        148299 => 2_000,
-        148304 => 2_500,
-        148322 => 2_000,
-        148327 => 1_500,
-        148332 => 3_000,
-        148337 => 2_500,
-        148347 => 1_500,
-        148348 => 2_000,
-        148349 => 2_000,
-        148354 => 2_000,
-        148362 => 1_500,
-        148378 => 3_000,
-        148389 => 3_000,
-        148393 => 3_000,
-        148400 => 3_000,
-        148402 => 1_500,
-        148407 => 1_500,
-        148423 => 1_500,
-        148429 => 3_000,
-        148430 => 3_000,
-        148431 => 2_500,
-        148433 => 3_000,
-        148437 => 2_000,
-        148476 => 2_000,
-        148521 => 2_000,
-        148522 => 2_500,
-        148527 => 3_000,
-        148532 => 3_000,
-        148534 => 3_000,
-        148538 => 3_000,
-        148539 => 1_500,
-        148541 => 2_500,
-        148543 => 1_500,
-        148544 => 3_000,
-        148545 => 2_500,
-        148546 => 1_500,
-        148547 => 3_000,
-        148548 => 3_000,
-        148553 => 2_000,
-        148558 => 3_000,
-        148560 => 3_000,
-        148562 => 2_000,
-        148572 => 3_000,
-        148573 => 2_000,
-        148575 => 1_500,
-        148578 => 3_000,
-        148582 => 2_000,
-        148583 => 2_500,
-        148589 => 2_000,
-        148590 => 1_500,
-        148596 => 3_000,
-        148601 => 2_000,
-        148616 => 3_000,
-        148641 => 2_000,
-        148654 => 3_000,
-        148655 => 3_000,
-        148684 => 2_000,
-        148692 => 2_000,
-        148694 => 2_000,
-        148703 => 1_500,
-        148706 => 3_000,
-        148712 => 1_500,
-        148722 => 2_000,
-        148728 => 1_500,
-        148737 => 2_000,
-        148768 => 2_000,
-        148771 => 3_000,
-        148772 => 2_000,
-        148790 => 3_000,
-        148792 => 3_000,
-        148793 => 3_000,
-        148826 => 2_000,
-        148830 => 2_500,
-        148835 => 1_500,
-        148847 => 2_000,
-        148886 => 2_000,
-        148901 => 2_000,
-        148902 => 2_000,
-        148921 => 1_500,
-        148932 => 1_500,
-        148936 => 3_000,
-        148965 => 3_000,
-        149009 => 2_000,
-        149011 => 3_000,
-        149031 => 3_000,
-        149032 => 3_000,
-        149047 => 2_000,
-        149151 => 3_000,
-        149155 => 2_000,
-        149162 => 2_000,
-        149197 => 2_000,
-        149205 => 2_500,
-        149235 => 3_000,
-        149299 => 2_500,
-        149328 => 1_500,
-        149380 => 2_500,
-        149396 => 2_000,
-        149429 => 3_000,
-        149430 => 3_000,
-        149431 => 1_500,
-        149432 => 2_000,
-        149433 => 2_000,
-        149436 => 2_000
-      }.freeze
-    end
-  end
-end
