--- app/models/levelling_up_premium_payments/eligibility_answers_presenter.rb
@@ -1,4 +1,0 @@
-module LevellingUpPremiumPayments
-  class EligibilityAnswersPresenter < Policies::EarlyCareerPayments::EligibilityAnswersPresenter
-  end
-end
