+++ app/models/policies/levelling_up_premium_payments/admin_tasks_presenter.rb
@@ -1,0 +1,13 @@
+module Policies
+  module LevellingUpPremiumPayments
+    class AdminTasksPresenter
+      extend Forwardable
+
+      def initialize(claim)
+        @existing_early_career_payments_admin_tasks_presenter = Policies::EarlyCareerPayments::AdminTasksPresenter.new(claim)
+      end
+
+      def_delegators :@existing_early_career_payments_admin_tasks_presenter, :employment, :identity_confirmation, :census_subjects_taught, :qualifications, :display_school
+    end
+  end
+end
