+++ app/models/policies/levelling_up_premium_payments/eligibility_answers_presenter.rb
@@ -1,0 +1,6 @@
+module Policies
+  module LevellingUpPremiumPayments
+    class EligibilityAnswersPresenter < Policies::EarlyCareerPayments::EligibilityAnswersPresenter
+    end
+  end
+end
