+++ app/models/policies/levelling_up_premium_payments/slug_sequence.rb
@@ -1,0 +1,6 @@
+module Policies
+  module LevellingUpPremiumPayments
+    class SlugSequence < Policies::EarlyCareerPayments::SlugSequence
+    end
+  end
+end
